requirements.txt
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.1
spacy==3.7.2
python-dateutil==2.8.2
uuid==1.30 